requirements.txt
google-auth==2.18.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.80.0
pandas==1.5.3
librosa==0.10.0
torch==1.13.1
transformers==4.24.0
#pip install google-cloud-speech
#pip install azure-cognitiveservices-speech